# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MetaSRS Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all classes
      parameters:
        - name: teacherId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get a class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Updated class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Class"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /classes/{id}/enroll:
    post:
      operationId: enrollStudent
      tags: [classes]
      summary: Enroll a student in a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollStudentBody"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /classes/{id}/students:
    get:
      operationId: listClassStudents
      tags: [classes]
      summary: List students enrolled in a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of enrolled students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentEnrollment"
  /students/{studentId}/classes:
    get:
      operationId: listStudentClasses
      tags: [classes]
      summary: List classes a student is enrolled in
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Class"

  # ─── Decks ───────────────────────────────────────────────────────────────────
  /decks:
    get:
      operationId: listDecks
      tags: [decks]
      summary: List decks
      parameters:
        - name: classId
          in: query
          required: false
          schema:
            type: integer
        - name: teacherId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deck"
    post:
      operationId: createDeck
      tags: [decks]
      summary: Create a deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeckBody"
      responses:
        "201":
          description: Created deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
  /decks/{id}:
    get:
      operationId: getDeck
      tags: [decks]
      summary: Get a deck by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDeck
      tags: [decks]
      summary: Update a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeckBody"
      responses:
        "200":
          description: Updated deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deck"
    delete:
      operationId: deleteDeck
      tags: [decks]
      summary: Delete a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Cards ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Users ─────────────────────────────────────────────────────────────────
...
    Class:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        subject:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: ["string", "null"]
        enrollmentCount:
          type: integer
        deckCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subject
        - teacherId
        - enrollmentCount
        - deckCount
        - createdAt
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        subject:
          type: string
        teacherId:
          type: integer
      required:
        - name
        - subject
        - teacherId
    UpdateClassBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        subject:
          type: string
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        classId:
          type: integer
        studentId:
          type: integer
        enrolledAt:
          type: string
          format: date-time
      required:
        - id
        - classId
        - studentId
        - enrolledAt
    EnrollStudentBody:
      type: object
      properties:
        studentId:
          type: integer
      required:
        - studentId
    StudentEnrollment:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: string
        enrolledAt:
          type: string
          format: date-time
        totalReviews:
          type: integer
        averageRetention:
          type: ["number", "null"]
        lastReviewedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - studentId
        - studentName
        - studentEmail
        - enrolledAt
        - totalReviews

    # ─── Decks ─────────────────────────────────────────────────────────────────
    Deck:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        teacherId:
          type: integer
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - teacherId
        - cardCount
        - createdAt
        - updatedAt
    CreateDeckBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        classId:
          type: integer
        teacherId:
          type: integer
      required:
        - name
        - teacherId
    UpdateDeckBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        classId:
          type: integer

    # ─── Cards ─────────────────────────────────────────────────────────────────
...