# lib/api-zod/src/generated/types/createDeckBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MetaSRS Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDeckBody {
  name: string;
  description?: string;
  classId?: number;
  teacherId: number;
}