# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MetaSRS Learning Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  classId: number;
  studentId: number;
  enrolledAt: Date;
}